[General] -- Hook. Now supports temporary flag for not saving any of the generated hook components. -- Preferences. Added user options for disableHooks and temporaryHooks. -- FileManager. Now initializes at editor start. Removed setup bool and all checks against it. -- FileManager. Now uses FileSystemWatchers per directory rather than reliance on unity events.  While this does cause a decent amount of cpu usage in the editor due to lack of rate throttling, it does occur in a unique thread.  May replace with multiple, former implementations. -- FileData. Folders should no longer have any data in fullName field. -- FileData. Added bool on Delete for deleting just the cache information.
[Extensions]
-- Array. Added Unshift() method for adding an item to the beginning of an array.
-- Object. Added Copy() method as a shorthand for a new instance and UseVariables call.
-- ObjectReflection. CallMethod() now properly scans all active assemblies.
-- ObjectReflection. UseVariables now returns the value afterwards for better inline usage.

[Themes]
-- Now tracks "s_Defaults" container when dumping GUISkin as well as clearing cache.
-- Style groups are now set if an Activator.CreateInstance was successful.
-- Renamed Parse methods to Import for better clarity. Overloaded to accept single path (where possible).
-- Renamed Save methods to Export for better clarity.
-- ThemeFont. Removed Copy method due to new ObjectExtension Copy extension.
-- ThemeFontset. Content offset are now added rather than set (fixes prefab icon issue)
-- ThemeUtility. SyncStyle now offers flipPattern parameter for determining which is the base theme and which is the sub.
-- ThemeWindow. Now handles actual Step/Update calls for better consistency between scene/mode changes.

[ThemeSystem]
-- Split window operations into ShowWindow() method.
-- Former static constructor contents now represented as Step() method.
-- Staggered needsRebuild operation so that two full OnGUI passes are done (giving enough time for other elements to generate styles).
-- The needsRefresh and needsRebuild bools are now set through Refresh and Rebuild methods.
-- Removed liveEdit hotkey due to accidental triggering and theme overwriting.
-- Missing fonts in fontsets will now correctly display a message rather than breaking the preferences window.

[Simplicity]
-- Removed JollyLodger font and Spooky fontset/font due to readability issues.
-- Fixed SquaresBoldFree naming within the Terminal fontset.
-- Now uses AnnotationWindow/BuildPlayerWindow skins for unique element styling/overriding.
-- Toggle and popup now display hover/focus states.
-- Profiler window styles improved.
-- Ordered List styles improved.
-- Reorderable List styles improved.
-- Toolbar separation with toggled buttons should now be more notable.
-- Window close, minimize, and maximize now have correctly styled images.
-- Minus/Plus as well as boxed inverted forms now have correctly styled images.
-- Color picker sliders now have correctly styled images.
-- Color field now has correctly styled images.

+++ Codebase/Systems/File/FileManager.cs
@@ -5,24 +5,24 @@ using System.IO;
 using UnityEngine;
 using UnityObject = UnityEngine.Object;
 namespace Zios{
-	using Events;
 	using Containers;
 	#if UNITY_EDITOR
 	using UnityEditor;
 	#endif
+	[InitializeOnLoad]
 	public static class FileManager{
 		private static string path;
-		private static bool setup;
 		private static bool debug = false;
 		private static bool clock = false;
 		private static bool fullScan = true;
-		private static DateTime pathModifyTime;
+		private static Dictionary<string,FileSystemWatcher> watchers = new Dictionary<string,FileSystemWatcher>();
 		public static Dictionary<string,List<FileData>> filesByPath = new Dictionary<string,List<FileData>>(StringComparer.InvariantCultureIgnoreCase);
 		public static Dictionary<string,List<FileData>> filesByType = new Dictionary<string,List<FileData>>(StringComparer.InvariantCultureIgnoreCase);
 		public static Dictionary<string,FileData> folders = new Dictionary<string,FileData>(StringComparer.InvariantCultureIgnoreCase);
 		public static Dictionary<string,FileData[]> cache = new Dictionary<string,FileData[]>();
 		public static Dictionary<UnityObject,object> assets = new Dictionary<UnityObject,object>();
 		public static Hierarchy<Type,string,string,UnityObject> namedAssets = new Hierarchy<Type,string,string,UnityObject>();
+		static FileManager(){FileManager.Refresh();}
 		//===============
 		// Storage
 		//===============
@@ -97,14 +97,18 @@ namespace Zios{
 		//===============
 		// Setup
 		//===============
-		public static void Monitor(){
-			var time = Time.realtimeSinceStartup;
-			var modifyTime = Directory.GetLastWriteTime(FileManager.path);
-			if(FileManager.pathModifyTime != modifyTime){
-				FileManager.pathModifyTime = modifyTime;
-				FileManager.Refresh();
-				if(FileManager.clock){Debug.Log("[FileManager] : Monitor sync complete -- " + (Time.realtimeSinceStartup-time) + " seconds.");}
+		public static void FileChanged(object source,FileSystemEventArgs data){
+			if(data.FullPath.ContainsAny(".svn","~",".meta","unitytemp","unitylock")){return;}
+			Utility.DelayCall(()=>FileManager.Refresh());
+			/*var path = Path.GetFullPath(data.FullPath).Replace("\\","/");
+			if(data.ChangeType == WatcherChangeTypes.Deleted){
+				FileManager.cache["!"+path.ToLower()][0].Delete(true);
 			}
+			if(data.ChangeType == WatcherChangeTypes.Created){
+				var fileData = new FileData(path);
+				fileData.isFolder = fileData.name.IsEmpty();
+				fileData.BuildCache();
+			}*/
 		}
 		public static void Refresh(){
 			var time = Time.realtimeSinceStartup;
@@ -114,8 +118,8 @@ namespace Zios{
 			FileManager.folders.Clear();
 			FileManager.cache.Clear();
 			FileManager.path = Application.dataPath.GetDirectory();
-			FileManager.pathModifyTime = Directory.GetLastWriteTime(FileManager.path);
 			FileManager.Scan(FileManager.path);
+			FileManager.Scan(FileManager.path+"/Temp",true);
 			if(!Application.isPlaying){
 				if(FileManager.fullScan){FileManager.Scan(Application.dataPath,true);}
 				FileManager.Save();
@@ -123,32 +127,38 @@ namespace Zios{
 			else{
 				FileManager.Load();
 			}
-			FileManager.setup = true;
-			Event.Add("On Editor Update",FileManager.Monitor).SetPermanent();
-			Event.Add("On Asset Changed",FileManager.Refresh).SetPermanent();
 			if(FileManager.clock){Debug.Log("[FileManager] : Refresh complete -- " + (Time.realtimeSinceStartup-time) + " seconds.");}
 		}
 		public static void Scan(string directory,bool deep=false){
+			if(!Directory.Exists(directory)){return;}
 			string[] fileEntries = Directory.GetFiles(directory);
 			string[] folderEntries = Directory.GetDirectories(directory);
+			FileManager.Monitor(directory);
 			FileManager.filesByPath.AddNew(directory);
 			foreach(string filePath in fileEntries){
-				if(filePath.Contains(".meta")){continue;}
+				if(filePath.ContainsAny(".meta","unitytemp","unitylock")){continue;}
 				var path = filePath.Replace("\\","/");
 				new FileData(path).BuildCache();
 			}
 			foreach(string folderPath in folderEntries){
-				if(folderPath.ContainsAny(".svn","~","Temp")){continue;}
+				if(folderPath.ContainsAny(".svn","~")){continue;}
 				var path = folderPath.Replace("\\","/");
 				new FileData(path,true).BuildCache();
 				if(deep){FileManager.Scan(path,true);}
+			}
+		}
+		public static void Monitor(string path){
+			if(!FileManager.watchers.ContainsKey(path)){
+				var watcher = FileManager.watchers[path] = new FileSystemWatcher(path);
+				watcher.Created += new FileSystemEventHandler(FileManager.FileChanged);
+				watcher.Deleted += new FileSystemEventHandler(FileManager.FileChanged);
+				watcher.EnableRaisingEvents = true;
 			}
 		}
 		//===============
 		// Primary
 		//===============
 		public static FileData[] FindAll(string name,bool showWarnings=true,bool firstOnly=false){
-			if(!FileManager.setup){FileManager.Refresh();}
 			var time = Time.realtimeSinceStartup;
 			if(name == "" && showWarnings){
 				Debug.LogWarning("[FileManager] No path given for search.");
@@ -212,7 +222,6 @@ namespace Zios{
 			}
 		}
 		public static string GetPath(UnityObject target,bool relative=true){
-			if(!FileManager.setup){FileManager.Refresh();}
 			#if UNITY_EDITOR
 			if(Application.isEditor){
 				string assetPath = AssetDatabase.GetAssetPath(target);
@@ -225,7 +234,6 @@ namespace Zios{
 		public static T GetAsset<T>(UnityObject target){
 			#if UNITY_EDITOR
 			if(Application.isEditor){
-				if(!FileManager.setup){FileManager.Refresh();}
 				if(!FileManager.assets.ContainsKey(target)){
 					string assetPath = FileManager.GetPath(target);
 					object asset = AssetDatabase.LoadAssetAtPath(assetPath,typeof(T));
@@ -324,9 +332,9 @@ namespace Zios{
 		public FileData(string path,bool isFolder=false){
 			this.path = path;
 			this.directory = path.GetDirectory();
-			this.extension = path.GetFileExtension();
 			this.name = path.GetFileName();
-			this.fullName = this.name + "." + this.extension;
+			this.extension = isFolder ? "" : path.GetFileExtension();
+			this.fullName = isFolder ? this.name : this.name + "." + this.extension;
 			this.isFolder = isFolder;
 		}
 		public void BuildCache(){
@@ -345,7 +353,7 @@ namespace Zios{
 		public void WriteText(string contents){
 			File.WriteAllText(this.path,contents);
 		}
-		public void Delete(){
+		public void Delete(bool cacheOnly=false){
 			foreach(var item in FileManager.cache.Copy()){
 				if(item.Value.Contains(this)){
 					FileManager.cache[item.Key] = item.Value.Remove(this);
@@ -355,12 +363,12 @@ namespace Zios{
 				}
 			}
 			if(!this.isFolder){
-				File.Delete(this.path);
+				if(!cacheOnly){File.Delete(this.path);}
 				FileManager.filesByType[this.extension].Remove(this);
-				FileManager.filesByPath[this.path].Remove(this);
+				FileManager.filesByPath[this.directory].Remove(this);
 				return;
 			}
-			Directory.Delete(this.path);
+			if(!cacheOnly){Directory.Delete(this.path);}
 			FileManager.folders.Remove(this.path);
 		}
 		public void MarkDirty(){File.SetLastWriteTime(this.path,DateTime.Now);}
